+++ src/main.ts
@@ -1,8 +1,7 @@
-import { ConfigService } from '@nestjs/config';
+import { ConfigService, ConfigType } from '@nestjs/config';
 import { NestFactory } from '@nestjs/core';
 import { SwaggerModule } from '@nestjs/swagger';
 import { AppModule } from './app.module';
-import { corsConfig } from './config/cors.config';
 import {
   swaggerConfig,
   swaggerOptions,
@@ -10,22 +9,23 @@ import {
 } from './config/swagger.config';
 import { Logger } from '@nestjs/common';
 import { ResponseLoggingInterceptor } from './log/response.logging.interceptor';
+import { MonitoringExternalModule } from './monitoring/monitoring.external-module';
+import { AppConfig } from './config/app.config';
+import { CorsConfig } from './config/cors.config';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
 
   const configService = app.get(ConfigService);
+  const appConfig = configService.get<ConfigType<AppConfig>>('application');
+  const corsConfig = configService.get<ConfigType<CorsConfig>>('cors');
 
   // Set up CORS
-  const origin = configService.get<string>('CORS_ALLOWED_ORIGINS').split(',');
   Logger.log(
-    `CORS enabled. The following origins will be allowed: '${origin.join(
-      "', '",
-    )}'.`,
-    'NestApplication',
+    `CORS enabled. The following origins will be allowed: '${corsConfig.origin}'.`,
+    'Bootstrap',
   );
   app.enableCors({
-    origin,
     ...corsConfig,
   });
 
@@ -37,8 +37,17 @@ async function bootstrap() {
 
   app.useGlobalInterceptors(new ResponseLoggingInterceptor());
 
-  const port = configService.get<number>('APP_PORT');
+  const { port } = appConfig;
   await app.listen(port);
-  Logger.log(`Application is listening on port ${port}.`, 'NestApplication');
+  Logger.log(`Application is listening on port ${port}.`, 'Bootstrap');
+
+  // Start monitoring process
+  const mon = await NestFactory.create(MonitoringExternalModule);
+  const { metricsPort } = appConfig;
+  await mon.listen(metricsPort);
+  Logger.log(
+    `Metrics are available at '/metrics' and port ${metricsPort}.`,
+    'Bootstrap',
+  );
 }
 bootstrap();